extend keypair generation feature - add feature to save keypair (configurable) - add feature to configure the keypair generation source

+++ ca-mgmt-client/src/main/java/org/xipki/ca/mgmt/db/port/DbPortWorker.java
@@ -88,7 +88,8 @@ public abstract class DbPortWorker extends DbWorker {
   }
 
   protected void decrypt(String dir) throws IOException {
-    ZipFile zipFile = new ZipFile(new File(dir, "encrypted/main.zip"), password);
+    File mainFile = new File(dir, "encrypted/main.zip");
+    ZipFile zipFile = new ZipFile(mainFile, password);
 
     boolean alreadyUnzipped = false;
     for (FileHeader fh : zipFile.getFileHeaders()) {
@@ -100,11 +101,9 @@ public abstract class DbPortWorker extends DbWorker {
       break;
     }
 
-    if (alreadyUnzipped) {
-      return;
+    if (!alreadyUnzipped) {
+      zipFile.extractAll(dir);
     }
-
-    zipFile.extractAll(dir);
   }
 
   protected void deleteDecryptedFiles(String dir) throws IOException {
@@ -113,25 +112,30 @@ public abstract class DbPortWorker extends DbWorker {
       return;
     }
 
+    File mainFile = new File(dir, "encrypted");
+    if (!mainFile.exists()) {
+      return;
+    }
+
     for (File f : files) {
       // delete all files and sub-dirs except 'encrypted'.
-        if (f.getName().equals("encrypted")) {
-          continue;
-        }
+      if (f.getName().equals("encrypted")) {
+        continue;
+      }
 
-        List<File> failedList = new LinkedList<>();
-        if (f.isFile()) {
-          if (!IoUtil.deleteFile(f)) {
-            failedList.add(f);
-          }
-        } else if (f.isDirectory()) {
-          if (!IoUtil.deleteDir(f)) {
-            failedList.add(f);
-          }
+      List<File> failedList = new LinkedList<>();
+      if (f.isFile()) {
+        if (!IoUtil.deleteFile(f)) {
+          failedList.add(f);
         }
+      } else if (f.isDirectory()) {
+        if (!IoUtil.deleteDir(f)) {
+          failedList.add(f);
+        }
+      }
 
-        if (!failedList.isEmpty()) {
-          LOG.error("error deleting files & folders: {}", failedList);
+      if (!failedList.isEmpty()) {
+        LOG.error("error deleting files & folders: {}", failedList);
       }
     }
   }
